refactor: 비동기 로직 분리
