+++ src/main/java/org/sopt/carena/healthreport/application/service/CreateHealthReportService.java
@@ -2,15 +2,12 @@ package org.sopt.carena.healthreport.application.service;
 
 import java.util.concurrent.ExecutorService;
 
-import org.sopt.carena.diet.application.port.out.EmbeddingPort;
-import org.sopt.carena.diet.exception.embedding.EmbeddingFailedException;
 import org.sopt.carena.healthreport.application.converter.HealthReportEmbeddingConverter;
 import org.sopt.carena.healthreport.application.dto.command.CreateHealthReportCommand;
 import org.sopt.carena.healthreport.application.port.in.CreateHealthReportUseCase;
-import org.sopt.carena.healthreport.application.port.out.HealthReportEmbeddingPersistencePort;
+import org.sopt.carena.healthreport.application.port.in.SaveHealthReportEmbeddingUseCase;
 import org.sopt.carena.healthreport.application.port.out.HealthReportPersistencePort;
 import org.sopt.carena.healthreport.domain.HealthReport;
-import org.sopt.carena.healthreport.domain.HealthReportEmbedding;
 import org.sopt.carena.healthreport.exception.healthreport.HealthReportAlreadyExistsException;
 import org.sopt.carena.member.application.port.in.HealthScoreUseCase;
 import org.sopt.carena.member.application.port.out.MemberPersistencePort;
@@ -18,9 +15,6 @@ import org.sopt.carena.member.domain.Member;
 
 import org.sopt.carena.member.exception.jwt.MemberNotFoundException;
 import org.sopt.carena.recommend.application.port.in.CreateRecommendedMealUseCase;
-import org.springframework.retry.annotation.Backoff;
-import org.springframework.retry.annotation.Recover;
-import org.springframework.retry.annotation.Retryable;
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
@@ -33,63 +27,31 @@ public class CreateHealthReportService implements CreateHealthReportUseCase {
 	private final CreateRecommendedMealUseCase createRecommendedMealUseCase;
 	private final HealthReportPersistencePort healthReportPersistencePort;
 	private final MemberPersistencePort memberPersistencePort;
-	private final HealthReportEmbeddingPersistencePort healthReportEmbeddingPersistencePort;
-	private final EmbeddingPort embeddingPort;
-	private final ExecutorService virtualExecutorService;
 	private final HealthScoreUseCase healthScoreUseCase;
+	private final SaveHealthReportEmbeddingUseCase saveHealthReportEmbeddingUseCase;
+	private final ExecutorService virtualExecutorService;
 
-	public void createHealthReport(final CreateHealthReportCommand commend) {
-		Member member = memberPersistencePort.getMemberById(commend.memberId())
+	public void createHealthReport(final CreateHealthReportCommand command) {
+		Member member = memberPersistencePort.getMemberById(command.memberId())
 				.orElseThrow(MemberNotFoundException::new);
 
 		if (healthReportPersistencePort.existsByMemberIdAndHealthCheckDate(
-				commend.memberId(),
-				commend.healthCheckDate()
+				command.memberId(),
+				command.healthCheckDate()
 		)) {
 			throw new HealthReportAlreadyExistsException();
 		}
 
 		HealthReport healthReport = healthReportPersistencePort
-				.saveHealthReport(HealthReport.create(commend, member.getGender()));
+				.saveHealthReport(HealthReport.create(command, member.getGender()));
 
-        String embeddingText = HealthReportEmbeddingConverter.toEmbeddingText(healthReport);
+		String embeddingText = HealthReportEmbeddingConverter.toEmbeddingText(healthReport);
 
 		//점수 계산 -> 여기선 멤버의  Usecase호출 / member에서 점수 계산~~~
 		healthScoreUseCase.updateMemberScore(member, healthReport);
 
-        virtualExecutorService.submit(() -> embeddingAndSave(embeddingText, member, healthReport));
+		virtualExecutorService.submit(
+				() -> saveHealthReportEmbeddingUseCase.embeddingAndSave(embeddingText, member, healthReport));
 		createRecommendedMealUseCase.saveRagResult(member.getId());
-	}
-
-	@Retryable(
-			retryFor = EmbeddingFailedException.class,
-			maxAttempts = 3,
-			backoff = @Backoff(delay = 5000, multiplier = 2),
-			recover = "recoverEmbedding"
-	)
-    public void embeddingAndSave(final String embeddingText, final Member member, final HealthReport healthReport) {
-        // 임베딩 호출
-        float[] embedding = embeddingPort.embed(embeddingText).vector();
-
-        // 임베딩 결과 도매인 생성
-        HealthReportEmbedding healthReportEmbedding = HealthReportEmbedding.builder()
-                .embeddingText(embeddingText)
-                .embedding(embedding)
-                .memberId(member.getId())
-                .healthReportId(healthReport.getId())
-                .build();
-
-        // 임베딩 결과 저장
-        healthReportEmbeddingPersistencePort.saveHealthReportEmbedding(healthReportEmbedding);
-    }
-
-	@Recover
-	public void recoverEmbedding(
-			final EmbeddingFailedException e,
-			final String embeddingText,
-			final Member member,
-			final HealthReport healthReport
-	){
-		log.error("임베딩 실패: {}",e.getMessage());
 	}
 }
